requirements.txt
fastapi
uvicorn
python-multipart
pydantic
numpy
pypdf
python-dotenv
google-generativeai
sentence-transformers
qdrant-client
boto3
